requirements.txt
requests
supabase
httpx
python-dotenv
fastapi
gunicorn
rapidfuzz
google-genai
lxml
asyncpg
google-cloud-aiplatform
uvicorn[standard]
filelock
python-dateutil
google-cloud-tasks
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
Jinja2
python-multipart
google-cloud-storage
google-cloud-run
geopy
upstash-redis
upstash-search
google-cloud-bigquery
